add qbit api methods, split downloaders
